+++ flowman-tools/src/main/scala/com/dimajix/flowman/tools/exec/sql/SqlCommand.scala
@@ -31,6 +31,7 @@ import com.dimajix.flowman.model.Mapping
 import com.dimajix.flowman.model.Project
 import com.dimajix.flowman.spec.mapping.SqlMapping
 import com.dimajix.flowman.tools.exec.Command
+import com.dimajix.flowman.util.ConsoleUtils
 
 
 class SqlCommand extends Command {
@@ -51,14 +52,7 @@ class SqlCommand extends Command {
         Try {
             val executor = session.execution
             val df = executor.instantiate(mapping, "main")
-            if (csv) {
-                val result = df.limit(limit).collect()
-                println(df.columns.mkString(","))
-                result.foreach(record => println(record.mkString(",")))
-            }
-            else {
-                df.show(limit)
-            }
+            ConsoleUtils.showDataFrame(df, limit, csv)
             true
         } match {
             case Failure(ex) =>