+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 * Implement new configurable `Executor` class for executing build targets.
 * Add build profile for Spark 3.1.x
 * Update ScalaTest to 3.2.5 - watch your unittests for changed ScalaTest API!
+* Add new `case` mapping
+* Add new `--dry-run` command line option
 
 
 # Version 0.14.2 - 2020-10-12